requirements.txt
paho-mqtt
requests
fastapi
uvicorn